Make mtp:: a backend-neutral API behind an MtpBackend seam
Phase 1 of the Windows WPD work: the high-level `mtp::` layer becomes
backend-neutral so PTP-over-USB and (later) WPD-over-COM can be two
implementations of one API. No behavior change on USB/virtual/mock — they all
ride the new `UsbBackend`, which wraps today's `PtpSession` logic unchanged.

- Introduce the `pub(crate)` `MtpBackend` trait (`mtp::backend`) in neutral
  vocabulary, with `UsbBackend` as the sole implementation. `MtpDevice`/`Storage`
  are now thin façades over `Box<dyn MtpBackend>`; all device-quirk handling
  (root-listing fast path, Android/Samsung/Fuji fallbacks, >4 GB size
  resolution, SIC cancel, recovery, the upload partial-handle contract) lives in
  `UsbBackend`, converting PTP↔neutral only at the boundary.
- Split the error model: the rich low-level error is renamed `Error` → `PtpError`
  (low-level `UploadError` → `PtpUploadError`); the crate-root `Error` is now the
  neutral `mtp::Error`, with `From<PtpError>` mapping response codes/USB faults
  into backend-agnostic variants (`StaleHandle`, `AccessDenied`, `Unsupported`,
  `Busy`, `StorageFull`, …). `ptp::` keeps `PtpError` for camera/protocol users.
- Opaque `u64` `ObjectHandle`/`StorageId` tokens, neutral `ObjectInfo`/
  `ObjectFormat`/`DeviceInfo`/`StorageInfo`/`DateTime`/`DeviceEvent`, and a
  neutral `Capabilities` (replaces `is_android`/`supports_*`; thin
  `supports_rename()`/`supports_upload()` kept as conveniences). Drop
  `MtpDevice::session()` — raw PTP access stays honest via `ptp::PtpDevice`.
- Consolidate downloads into one `ByteRange { Full, From, Range }` primitive:
  streaming `Storage::download(handle, range)`, session-releasing
  `download_windowed(handle, range, window_size)`, and buffered
  `download_to_vec` / `read_range` / `thumbnail`. `FileDownload`/`ObjectListing`
  wrap boxed backend streams.
- Add a backend-conformance test suite (`tests/conformance.rs`) against the
  public neutral API on the virtual device: round-trip, ranged/windowed
  download, fs ops, recursive listing, progress, and cancel/drop. Port the
  storage-quirk regression tests into `UsbBackend`.
- Document the seam in AGENTS.md.

+++ crates/mtp-rs/src/mtp/event.rs
@@ -1,6 +1,7 @@
 //! Device events.
 
-use crate::ptp::{EventCode, EventContainer, ObjectHandle, StorageId};
+use crate::mtp::{ObjectHandle, StorageId};
+use crate::ptp::{EventCode, EventContainer};
 
 /// Events from an MTP device.
 #[derive(Debug, Clone)]
@@ -62,22 +63,22 @@ impl DeviceEvent {
     pub fn from_container(container: &EventContainer) -> Self {
         match container.code {
             EventCode::ObjectAdded => DeviceEvent::ObjectAdded {
-                handle: ObjectHandle(container.params[0]),
+                handle: ObjectHandle(u64::from(container.params[0])),
             },
             EventCode::ObjectRemoved => DeviceEvent::ObjectRemoved {
-                handle: ObjectHandle(container.params[0]),
+                handle: ObjectHandle(u64::from(container.params[0])),
             },
             EventCode::StoreAdded => DeviceEvent::StoreAdded {
-                storage_id: StorageId(container.params[0]),
+                storage_id: StorageId(u64::from(container.params[0])),
             },
             EventCode::StoreRemoved => DeviceEvent::StoreRemoved {
-                storage_id: StorageId(container.params[0]),
+                storage_id: StorageId(u64::from(container.params[0])),
             },
             EventCode::StorageInfoChanged => DeviceEvent::StorageInfoChanged {
-                storage_id: StorageId(container.params[0]),
+                storage_id: StorageId(u64::from(container.params[0])),
             },
             EventCode::ObjectInfoChanged => DeviceEvent::ObjectInfoChanged {
-                handle: ObjectHandle(container.params[0]),
+                handle: ObjectHandle(u64::from(container.params[0])),
             },
             EventCode::DeviceInfoChanged => DeviceEvent::DeviceInfoChanged,
             // All other codes (including Unknown and unhandled known codes like DevicePropChanged)
@@ -113,7 +114,7 @@ mod tests {
                 DeviceEvent::ObjectInfoChanged { handle } => handle,
                 _ => panic!("Unexpected event type"),
             };
-            assert_eq!(handle, ObjectHandle(expected_handle));
+            assert_eq!(handle, ObjectHandle(u64::from(expected_handle)));
         }
 
         // Events with storage ID param
@@ -134,7 +135,7 @@ mod tests {
                 DeviceEvent::StorageInfoChanged { storage_id } => storage_id,
                 _ => panic!("Unexpected event type"),
             };
-            assert_eq!(storage_id, StorageId(expected_id));
+            assert_eq!(storage_id, StorageId(u64::from(expected_id)));
         }
 
         // DeviceInfoChanged (no params)
